version_1/requirements-agent.txt
# Parity Agent Dependencies
# Install with: pip install -r parity_agent/requirements-agent.txt

# Core ML
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.14.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Configuration & storage
PyYAML>=6.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Testing
pytest>=7.4.0

# Utilities
tqdm>=4.65.0
